p29.24 refactor: thread ids.UserID everywhere
Typed-ID capstone: UserID is now a distinct type across store + web + sqlc +
CLI. Purely types, no behavior change.

Threaded:
- sqlc: User.ID; UsersVersion.EntityID + UserReportGrantsVersion.EntityID (audit
  PKs .ID stay int64); UserReportGrant.UserID; ImportBatch.UploadedBy +
  InsertImportBatchParams.UploadedBy; ExpenseReport.SubmitterID (+ version) +
  InsertExpenseReportParams.SubmitterID; every users.sql.go id param/return
  (GetUser/GetUserRow/UserByID id, CountOtherEnabledAdmins id, InsertUser return,
  InsertUserVersion.ID_2 entity_id, the six Set*Params.ID, UserIDByUsername
  return); report_groups.sql.go grant UserID params + InsertReportGrantVersion.
  UserID; ListExpenseReportsBySubmitter param; LastHeaderAccountForActor actorID
  param.
- store: Actor.ID + CurrentUser.ID + AdminUser.ID + Credentials.ID; systemUserID
  const; every user*/grant/submitter method param + insertUserVersion entityID;
  CreateUser/UserIDByUsername returns.
- web: parse to typed at the parseID seam (admin_user_detail, admin_users);
  render-only model fields stay int64 (int64(u.ID) at assignment, userDetailURL
  callers wrapped); submitterNameMap keyed map[ids.UserID]string; systemUserWebID.
- CLI/synth: ledgerimport systemActorID, devseed firstUser, demo user ids +
  createDemoUser typed.
- tests: mkUser/mkSubmitter/adminUserByUsername/env user fields + user-id helper
  params typed; generic testutil helpers (AssertVersioned/Grant, LatestVersion/
  GrantActor) stay int64 -- callers pass int64(userID).

Boundary discipline (Rule 15): the SESSION is an int64 storage boundary on BOTH
sides -- authenticate() returns int64(creds.ID) into scs Put, middleware wraps
ids.UserID(GetInt64(...)); test mintCookie Puts int64(userID). (scs Put takes
interface{}, so this is a RUNTIME seam the compiler cannot catch -- gob-encoding
a named type / GetInt64's int64 assertion would fail; verified by running the
web suite, not just building it.)

Deferred (audit-actor plumbing, its own final batch with ChangeID): Change.ActorID
+ InsertChange.ActorID stay int64 (one int64(actor.ID) wrap at the InsertChange
call); the changes/history ActorID read projections (transactions.sql.go,
store/history.go HistoryEntry.ActorID, web latestChangeActorByKind) stay int64.
The write-funnel changeID and *_versions.ID PKs stay int64 (ChangeID reserved).

+++ internal/store/reconciliations_test.go
@@ -367,7 +367,7 @@ func TestReconciliationReopenAudited(t *testing.T) {
 	if err := e.s.Reopen(reopenCtx, recon); err != nil {
 		t.Fatalf("reopen: %v", err)
 	}
-	if actor := testutil.LatestVersionActor(t, e.d, "reconciliations", int64(recon)); actor != reopener {
+	if actor := testutil.LatestVersionActor(t, e.d, "reconciliations", int64(recon)); actor != int64(reopener) {
 		t.Errorf("reopen version actor = %d, want %d", actor, reopener)
 	}
 }
